feat (auth): - Удален widget SugnUp - Содержимое  widget SugnUp перенесено в page - Создан новый widget Confirmation - Реализован запрос на повторное потверждение в widget Confirmation - Переделан компонент Infoform(Добавлены новые  пропсы) - Добавлена svg стрелка на кнопке в AgreementContent - Исправлены стили alert
