+++ app/auth/signup/page.module.css
@@ -1,0 +1,5 @@
+.page {
+  padding: 24px 0;
+  background-color: var(--color-dark-700);
+  height: 100vh;
+}
