+++ features/alert/alert.module.css
@@ -1,7 +1,7 @@
 .container {
   position: fixed;
   left: 20px;
-  bottom: 20px;
+  bottom: 60px;
   display: inline-block;
   align-content: center;
   min-width: 387px;
