+++ features/auth/ui/Register/InfoForm/InfoForm.tsx
@@ -1,3 +1,4 @@
+"use client"
 import Image, { StaticImageData } from "next/image"
 import s from "./InfoForm.module.css"
 import Button from "@/shared/ui/button/button"
@@ -6,6 +7,9 @@ import { TextField } from "@/shared/ui/text-field/text-field"
 import { useForm } from "react-hook-form"
 import { zodResolver } from "@hookform/resolvers/zod"
 import { SignUpInputs, signUpSchema } from "@/shared/lib/signUpSchema/signUpSchema"
+import { ResendingEmailType } from "@/features/auth/api/authApi"
+import { setAppError } from "@/app/model/appSlice"
+import { useAppDispatch } from "@/app/hooks/useAppDispatch"
 
 type Props = {
   title: string
@@ -14,9 +18,11 @@ type Props = {
   img: StaticImageData
   isInput: boolean
   href?: string
+  handleClick?: (prov: ResendingEmailType, reset: () => void) => void
 }
 
-export const InfoForm = ({ title, text, textBtn, isInput, img, href }: Props) => {
+export const InfoForm = ({ title, text, textBtn, isInput, img, href, handleClick }: Props) => {
+  const dispatch = useAppDispatch()
   // Валидация React-hook-form
   const {
     register,
@@ -28,8 +34,14 @@ export const InfoForm = ({ title, text, textBtn, isInput, img, href }: Props) =>
     defaultValues: { email: "" },
   })
 
-  const onSubmit = () => {
-    reset()
+  const onSubmit = (data: Partial<SignUpInputs>) => {
+    const email = data.email as string
+    if (email !== undefined) {
+      // @ts-expect-error
+      handleClick({ email, baseUrl: "http://localhost:3000/auth/signup" }, reset)
+    } else {
+      dispatch(setAppError({ error: "Email is required" }))
+    }
   }
 
   return (
@@ -45,7 +57,7 @@ export const InfoForm = ({ title, text, textBtn, isInput, img, href }: Props) =>
             type={"text"}
             {...register("email")}
             errorMessage={errors.email?.message}
-            // className={s.field} доделать
+            className={s.field}
           />
           <Button>{textBtn}</Button>
         </Form.Root>
