+++ features/auth/ui/Register/ModalWindow/ModalWindow.module.css
@@ -1,6 +1,6 @@
 .fon {
   position: absolute;
-  top:0;
+  top: 0;
   left: 0;
   background-color: var(--color-dark-700);
   height: 100%;
