+++ package.json
@@ -18,7 +18,7 @@
     "@hookform/resolvers": "^5.1.1",
     "@radix-ui/themes": "^3.2.1",
     "@reduxjs/toolkit": "^2.8.2",
-    "next": "15.3.5",
+    "next": "^15.4.5",
     "next-auth": "^4.24.11",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
